requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
gspread==6.1.2
google-auth==2.35.0
gunicorn
